recommenderAI/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0 